requirements_windows.txt
django==4.2
django-environ==0.10.0
psycopg2==2.9.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5